feat(context): refactor context observer subscription into a new class
